refactor: rename task-models catalog
