+++ src/app/features/private/pomodoro-session/components/pomodoro-session-accept-task-modal/pomodoro-session-accept-task-modal.component.ts
@@ -2,7 +2,7 @@ import { Component, EventEmitter, Input, Output } from '@angular/core';
 
 import { NzButtonModule } from 'ng-zorro-antd/button';
 import { NzModalModule } from 'ng-zorro-antd/modal';
-import { ActivityItemDetails } from '../../../../shared/models/task-models/activity.model';
+import { ActivityItemDetails } from '../../../../shared/models/task/activity.model';
 
 @Component({
   selector: 'app-pomdoro-accept-task-modal',
