+++ src/app/features/private/shared/pomodoro-form-modal/pomodoro-session-form-modal.ts
@@ -11,11 +11,11 @@ import { NzButtonModule } from 'ng-zorro-antd/button';
 import { NzModalModule } from 'ng-zorro-antd/modal';
 import { PomodoroFormComponent } from '../pomodoro-form/pomodoro-form.component';
 import { UserPomodoroApiService } from '../../../shared/services/tasks/user-pomodoro-api.service';
-import { PomodoroRequest } from '../../../shared/models/task-models/pomodoro-request';
+import { PomodoroRequest } from '../../../shared/models/task/pomodoro-request';
 import {
   ActivityItemDetails,
   ActivityType,
-} from '../../../shared/models/task-models/activity.model';
+} from '../../../shared/models/task/activity.model';
 import { NotificationService } from '../../../shared/services/notification-service/notification.service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
